requirements-streamlit.txt
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
streamlit==1.51.0
plotly==6.4.0
requests==2.32.3
pyarrow==19.0.0
python-dotenv==1.0.1
joblib==1.4.2
scipy==1.15.2
